+++ bitstring/__init__.py
@@ -18,6 +18,8 @@ pack -- Create a Bits object from a format string.
 Exceptions:
 
 ReadError -- Reading or peeking past the end of a bitstring. Subclasses ValueError.
+InterpretationError -- A bitstring's length has no interpretation as the dtype asked
+                       for. Subclasses AttributeError and ValueError.
 
 https://github.com/scott-griffiths/bitstring
 """
@@ -57,7 +59,7 @@ from .bitarray_ import BitArray
 from .reader import Reader
 from .methods import pack
 from .array_ import Array
-from .exceptions import ReadError
+from .exceptions import ReadError, InterpretationError
 from .dtypes import DtypeDefinition as _DtypeDefinition, dtype_register as _dtype_register, Dtype
 from typing import Literal as _Literal
 
@@ -296,4 +298,4 @@ if Bits.__doc__ is not None:
 if BitArray.__doc__ is not None:
     BitArray.__doc__ = BitArray.__doc__.replace('[GENERATED_PROPERTY_DESCRIPTIONS]', _property_docstring)
 __all__ = ['Reader', 'BitArray', 'Array',
-           'Bits', 'pack', 'ReadError', 'Dtype']
+           'Bits', 'pack', 'ReadError', 'InterpretationError', 'Dtype']